[Connector] Change publish_object() signature
The interface_index parameter relates to the corresponding xml
part from the introspection data passed to initialize().
This was convenient for the d-bus connector implementation, but
required xml parsing for other connector implementations.

This commit replaces the interface_index parameter with the
actual interface name to avoid the xml parsing task.

Signed-off-by: Regis Merlino <regis.merlino@intel.com>