Port to gupnp-1.2
The latest version of gupnp breaks backwards compatibility. Fortunately,
there are not many calls to gupnp functions in core, so we just needed
to bump the dependencies.

There is one deprecated gupnp_service_proxy_cancel_action call but
I am not sure how to change it without breaking Dleyna API.
Let's leave it for now.